tidy code
