+++ cl/cltypes/beacon_block.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
+	"github.com/ledgerwatch/erigon-lib/types/clonable"
 	"github.com/ledgerwatch/erigon-lib/types/ssz"
 
 	"github.com/ledgerwatch/erigon/cl/clparams"
@@ -17,6 +18,8 @@ var (
 	_ ssz2.SizedObjectSSZ = (*BeaconBody)(nil)
 	_ ssz2.SizedObjectSSZ = (*BeaconBlock)(nil)
 	_ ssz2.SizedObjectSSZ = (*SignedBeaconBlock)(nil)
+	_ ssz2.SizedObjectSSZ = (*DenebBeaconBlock)(nil)
+	_ ssz2.SizedObjectSSZ = (*DenebSignedBeaconBlock)(nil)
 )
 
 const (
@@ -378,4 +381,78 @@ func (b *BeaconBody) UnmarshalJSON(buf []byte) error {
 	b.ExecutionChanges = tmp.ExecutionChanges
 	b.BlobKzgCommitments = tmp.BlobKzgCommitments
 	return nil
+}
+
+type DenebBeaconBlock struct {
+	Block     *BeaconBlock              `json:"block"`
+	KZGProofs *solid.ListSSZ[*KZGProof] `json:"kzg_proofs"`
+	Blobs     *solid.ListSSZ[*Blob]     `json:"blobs"`
+}
+
+func NewDenebBeaconBlock(beaconCfg *clparams.BeaconChainConfig) *DenebBeaconBlock {
+	maxBlobsPerBlock := int(beaconCfg.MaxBlobsPerBlock)
+	b := &DenebBeaconBlock{
+		Block:     NewBeaconBlock(beaconCfg),
+		KZGProofs: solid.NewStaticListSSZ[*KZGProof](maxBlobsPerBlock, BYTES_KZG_PROOF),
+		Blobs:     solid.NewStaticListSSZ[*Blob](maxBlobsPerBlock, int(BYTES_PER_BLOB)),
+	}
+	return b
+}
+
+func (b *DenebBeaconBlock) EncodeSSZ(buf []byte) ([]byte, error) {
+	return ssz2.MarshalSSZ(buf, b.Block, b.KZGProofs, b.Blobs)
+}
+
+func (b *DenebBeaconBlock) DecodeSSZ(buf []byte, version int) error {
+	return ssz2.UnmarshalSSZ(buf, version, b.Block, b.KZGProofs, b.Blobs)
+}
+
+func (b *DenebBeaconBlock) EncodingSizeSSZ() int {
+	return b.Block.EncodingSizeSSZ() + b.KZGProofs.EncodingSizeSSZ() + b.Blobs.EncodingSizeSSZ()
+}
+
+func (b *DenebBeaconBlock) Clone() clonable.Clonable {
+	return &DenebBeaconBlock{}
+}
+
+func (b *DenebBeaconBlock) Static() bool {
+	// it's variable size
+	return false
+}
+
+type DenebSignedBeaconBlock struct {
+	SignedBlock *SignedBeaconBlock        `json:"signed_block"`
+	KZGProofs   *solid.ListSSZ[*KZGProof] `json:"kzg_proofs"`
+	Blobs       *solid.ListSSZ[*Blob]     `json:"blobs"`
+}
+
+func NewDenebSignedBeaconBlock(beaconCfg *clparams.BeaconChainConfig) *DenebSignedBeaconBlock {
+	maxBlobsPerBlock := int(beaconCfg.MaxBlobsPerBlock)
+	b := &DenebSignedBeaconBlock{
+		SignedBlock: NewSignedBeaconBlock(beaconCfg),
+		KZGProofs:   solid.NewStaticListSSZ[*KZGProof](maxBlobsPerBlock, BYTES_KZG_PROOF),
+		Blobs:       solid.NewStaticListSSZ[*Blob](maxBlobsPerBlock, int(BYTES_PER_BLOB)),
+	}
+	return b
+}
+
+func (b *DenebSignedBeaconBlock) EncodeSSZ(buf []byte) ([]byte, error) {
+	return ssz2.MarshalSSZ(buf, b.SignedBlock, b.KZGProofs, b.Blobs)
+}
+
+func (b *DenebSignedBeaconBlock) DecodeSSZ(buf []byte, version int) error {
+	return ssz2.UnmarshalSSZ(buf, version, b.SignedBlock, b.KZGProofs, b.Blobs)
+}
+
+func (b *DenebSignedBeaconBlock) EncodingSizeSSZ() int {
+	return b.SignedBlock.EncodingSizeSSZ() + b.KZGProofs.EncodingSizeSSZ() + b.Blobs.EncodingSizeSSZ()
+}
+
+func (b *DenebSignedBeaconBlock) Clone() clonable.Clonable {
+	return &DenebSignedBeaconBlock{}
+}
+
+func (b *DenebSignedBeaconBlock) Static() bool {
+	// it's variable size
+	return false
 }
